add single server upgrade test suite
