+++ release_tester/license_manager_tests/base/single_server_base.py
@@ -1,5 +1,5 @@
-"""License manager tests: active failover"""
-import requests
+"""License manager tests: single server"""
+import json
 
 # pylint: disable=import-error
 from arangodb.installers import RunProperties
@@ -10,57 +10,53 @@ from reporting.reporting_utils import step
 from test_suites_core.base_test_suite import run_before_suite
 
 
-class LicenseManagerAfoBaseTestSuite(LicenseManagerBaseTestSuite):
-    """License manager tests: active failover"""
+class LicenseManagerSingleServerBaseTestSuite(LicenseManagerBaseTestSuite):
+    """License manager tests: single server"""
 
     def get_default_instance_type(self):
         """get the instance type we should communicate with"""
-        return InstanceType.RESILIENT_SINGLE
+        return InstanceType.SINGLE
 
     @run_before_suite
-    def startup(self):
-        """prepare test environment to run license manager tests on AFO setup"""
-        self.start_afo()
+    def start(self):
+        """start a single server setup before running tests"""
+        self.start_single_server()
 
     @step
-    def start_afo(self):
-        """start an AFO setup"""
-        # pylint: disable=attribute-defined-outside-init
+    def start_single_server(self):
+        """start a single server setup"""
         self.runner = make_runner(
-            runner_type=RunnerType.ACTIVE_FAILOVER,
+            runner_type=RunnerType.SINGLE,
             abort_on_error=False,
+            installer_set=self.installer_set,
+            use_auto_certs=False,
             selenium_worker="none",
             selenium_driver_args=[],
-            installer_set=self.installer_set,
             runner_properties=RunProperties(
                 enterprise=True,
                 encryption_at_rest=False,
                 ssl=False,
             ),
-            use_auto_certs=False,
         )
         self.runner.starter_prepare_env()
         self.runner.starter_run()
         self.runner.finish_setup()
-        self.starter = self.runner.leader
+        self.runner.starter_instance.detect_arangosh_instances()
+        self.starter = self.runner.starter_instance
 
     def get_server_id(self):
-        """read server ids from agency"""
-        agents = self.get_agents()
-        agent_ids = []
-        for agent in agents:
-            agent_ids.append((agent.basedir / "data" / "UUID").read_text())
-        agent_ids.sort()
-        return "".join(agent_ids)
+        """read server id from data dir"""
+        datadir = self.starter.all_instances[0].basedir / "data"
+        server_file_content = json.load(open(datadir / "SERVER"))
+        server_id = server_file_content["serverId"]
+        return server_id
 
     # pylint: disable=redefined-builtin
     def set_license(self, license):
         """set new license"""
-        body = """[[{"/arango/.license":{"op":"set","new": """ + license + """}}]]"""
-        resp = self.runner.agency_get_leader_starter_instance().send_request(
-            InstanceType.AGENT,
-            requests.post,
-            "/_api/agency/write",
-            body,
-        )
-        assert 200 <= resp[0].status_code < 300, "Failed to write license to agency."
+        datadir = self.starter.all_instances[0].basedir / "data"
+        with open(datadir / ".license", "w") as license_file:
+            license_file.truncate()
+            license_file.write(license)
+        self.starter.terminate_instance()
+        self.starter.respawn_instance()
