+++ release_tester/license_manager_tests/upgrade/afo.py
@@ -11,18 +11,6 @@ from test_suites_core.base_test_suite import testcase
 class LicenseManagerAfoUpgradeTestSuite(LicenseManagerAfoBaseTestSuite, LicenseManagerUpgradeBaseTestSuite):
     """License manager tests: upgrade AFO"""
 
-    @step
-    def upgrade_afo(self):
-        """upgrade a AFO setup"""
-        self.new_installer.calculate_package_names()
-        self.new_installer.upgrade_server_package(self.old_installer)
-        self.new_installer.output_arangod_version()
-        self.new_installer.stop_service()
-        self.runner.cfg.set_directories(self.new_installer.cfg)
-        self.runner.new_cfg.set_directories(self.new_installer.cfg)
-        self.runner.upgrade_arangod_version()  # make sure to pass new version
-        self.old_installer.un_install_server_package_for_upgrade()
-
     @testcase
     def upgrade_when_license_is_expired(self):
         """Check that upgrade can be performed with expired license"""
@@ -39,7 +27,7 @@ class LicenseManagerAfoUpgradeTestSuite(LicenseManagerAfoBaseTestSuite, LicenseM
                 expect_to_fail=False,
             )
         self.expire_license()
-        self.upgrade_afo()
+        self.upgrade()
         with step("check that data is present after the upgrade"):
             try:
                 self.starter.arangosh.run_command(
