+++ release_tester/license_manager_tests/upgrade/dc2dc.py
@@ -12,18 +12,6 @@ from test_suites_core.base_test_suite import testcase, disable_for_windows
 class LicenseManagerDc2DcUpgradeTestSuite(LicenseManagerDc2DcBaseTestSuite, LicenseManagerUpgradeBaseTestSuite):
     """License manager tests: upgrade DC2DC setup"""
 
-    @step
-    def upgrade_dc2dc(self):
-        """upgrade a cluster setup"""
-        self.new_installer.calculate_package_names()
-        self.new_installer.upgrade_server_package(self.old_installer)
-        self.new_installer.output_arangod_version()
-        self.new_installer.stop_service()
-        self.runner.cfg.set_directories(self.new_installer.cfg)
-        self.runner.new_cfg.set_directories(self.new_installer.cfg)
-        self.runner.upgrade_arangod_version()  # make sure to pass new version
-        self.old_installer.un_install_server_package_for_upgrade()
-
     @testcase
     def upgrade_when_license_is_expired(self):
         """Check that upgrade can be performed with expired license"""
@@ -45,7 +33,7 @@ class LicenseManagerDc2DcUpgradeTestSuite(LicenseManagerDc2DcBaseTestSuite, Lice
                 expect_to_fail=False,
             )
         self.expire_license()
-        self.upgrade_dc2dc()
+        self.upgrade()
         with step("check that data is present after the upgrade"):
             try:
                 self.starter.arangosh.run_command(
