+++ release_tester/license_manager_tests/upgrade/leader_follower.py
@@ -13,18 +13,6 @@ class LicenseManagerLeaderFollowerUpgradeTestSuite(
 ):
     """License manager tests: upgrade leader-follower"""
 
-    @step
-    def upgrade_leader_follower(self):
-        """upgrade a leader-follower setup"""
-        self.new_installer.calculate_package_names()
-        self.new_installer.upgrade_server_package(self.old_installer)
-        self.new_installer.output_arangod_version()
-        self.new_installer.stop_service()
-        self.runner.cfg.set_directories(self.new_installer.cfg)
-        self.runner.new_cfg.set_directories(self.new_installer.cfg)
-        self.runner.upgrade_arangod_version()  # make sure to pass new version
-        self.old_installer.un_install_server_package_for_upgrade()
-
     @testcase
     def upgrade_when_license_is_expired(self):
         """Check that upgrade can be performed with expired license"""
@@ -41,7 +29,7 @@ class LicenseManagerLeaderFollowerUpgradeTestSuite(
                 expect_to_fail=False,
             )
         self.expire_license()
-        self.upgrade_leader_follower()
+        self.upgrade()
         with step("check that data is present after the upgrade"):
             try:
                 self.runner.leader_starter_instance.arangosh.run_command(
