+++ release_tester/license_manager_tests/upgrade/upgrade_test_suite.py
@@ -7,6 +7,7 @@ from license_manager_tests.upgrade.afo import LicenseManagerAfoUpgradeTestSuite
 from license_manager_tests.upgrade.cluster import LicenseManagerClusterUpgradeTestSuite
 from license_manager_tests.upgrade.dc2dc import LicenseManagerDc2DcUpgradeTestSuite
 from license_manager_tests.upgrade.leader_follower import LicenseManagerLeaderFollowerUpgradeTestSuite
+from license_manager_tests.upgrade.single_server import LicenseManagerSingleServerUpgradeTestSuite
 
 IS_WINDOWS = platform.win32_ver()[0] != ""
 
@@ -15,10 +16,9 @@ class UpgradeLicenseManagerTestSuite(LicenseManagerBaseTestSuite):
     """License manager test suite: Upgrade"""
 
     child_test_suites = [
+        LicenseManagerSingleServerUpgradeTestSuite,
         LicenseManagerLeaderFollowerUpgradeTestSuite,
         LicenseManagerClusterUpgradeTestSuite,
         LicenseManagerAfoUpgradeTestSuite,
+        LicenseManagerDc2DcUpgradeTestSuite,
     ]
-
-    if not IS_WINDOWS:
-        child_test_suites.append(LicenseManagerDc2DcUpgradeTestSuite)
